added an entire drag and drop system with a mouse

+++ src/functions/transform/transform.c
@@ -4,23 +4,23 @@
 #include "../genericsHeader.h"
 #include "../../components/componentsGenarics.h"
 
-void movePosition(Entity e, int x, int y)
+void movePosition(Entity e, float x, float y)
 {
     e->components.transform->position.x += x;
     e->components.transform->position.y += y;
 }
-void setPosition(Entity e, int x, int y)
+void setPosition(Entity e, float x, float y)
 {
     e->components.transform->position.x = x;
     e->components.transform->position.y = y;
 }
 
-void moveScale(Entity e, int x, int y)
+void moveScale(Entity e, float x, float y)
 {
     e->components.transform->scale.x += x;
     e->components.transform->scale.y += y;
 }
-void setScale(Entity e, int x, int y)
+void setScale(Entity e, float x, float y)
 {
     e->components.transform->scale.x = x;
     e->components.transform->scale.y = y;
