+++ src/utils/math/MMath.h
@@ -2,11 +2,11 @@
 
 struct Vector2
 {
-    int x;
-    int y;
+    float x;
+    float y;
 };
 
 struct Vector2;
-struct Vector2 NewVector2(int x, int y);
+struct Vector2 NewVector2(float x, float y);
 
 void PerpendicularLinePostulate(struct Vector2 a, struct Vector2 b, struct Vector2 p);
